Put the two select cases in run.go in their own files
